Added service codes for USPS Domestic and USPS International and other methods
Service codes now added to USPS Domestic and USPS International
Providers. FedEx and UPS providers as well as the USPS providers now
have a method called GetServiceCodes which will return them. This is a
way for systems to save these and then build logic on top of the rate
lookups to exclude certain rates from showing on their UIs/systems based
on a specific service code for a certain carrier.